* fixed a small issue in the info formater.
* deleted the old README and made a HTML version

* made a dummy command which cats the HTML version

* included the info and help command in the info.plist
  (if its on the wrong place please reorder, chirs)


as you maybe know, i dont speak english native.
so it would probably be better if someone else trys
to describe alle the svn commands.  i created a small
infastructure for the documentation, but feel free
to alter it / delete it if you want to write it in
another way.



git-svn-id: http://svn.textmate.org/trunk/Subversion.tmbundle@332 dfb7d73b-c2ec-0310-8fea-fb051d288c6d