• added grammar for 'svn-commit.tmp' files
This is modeled over the diff grammar and marks up the separator line plus added, deleted, and modified files (as markup.(inserted|changed|deleted).svn).

The main motivation behind this grammar though was scoping all the stuff below the separator as meta.scope.changed-files.svn so that (continuous) spell checking could be disabled for that part (seeing how I've fixed Check Spelling as You Type to respect scope specific settings (and not underline the word touched), so now I have this enabled all the time :) ).



git-svn-id: http://svn.textmate.org/trunk/Bundles/Subversion.tmbundle@1293 dfb7d73b-c2ec-0310-8fea-fb051d288c6d