Switch the status/update window to the standard TextMate style.
On the coding side, some minor cleanup and refactoring remains.

The CSS, however, could use some serious design work -- it now looks rather bland. I've added only the styles for highlighting the status column, which really only work in the 'bright' theme, and dropped everything else (because it conflicted with the colors in all of the themes).

git-svn-id: http://svn.textmate.org/trunk/Bundles/Subversion.tmbundle@5790 dfb7d73b-c2ec-0310-8fea-fb051d288c6d