• we append to revision_comment[revision_number] but incase we never see a /^r(\d+)/ line, this will be nil — see http://lists.macromates.com/pipermail/textmate/2008-March/024800.html
This fix only avoids the exception, incase we do not have a revision number, we have output from svn that we don’t expect, so there might be other problems, or it might just be, that svn prints a banner before the actual log, e.g. as part of the underlying ssh login process.

It might be that we should ignore stderr rather than merge it with stdout, and long-term we should switch to the --xml option of svn.


git-svn-id: http://svn.textmate.org/trunk/Bundles/Subversion.tmbundle@9220 dfb7d73b-c2ec-0310-8fea-fb051d288c6d