• Class that represents what `svn status` indicates.
The status information is calculated by parsing XML output of svn, opposed to pushing the normal output through REs, which should hopefully keep us safe for any language that svn is running in.

At the moment this is only suitable for using with the commit command. This will be used for *all* status getting when the revamp is done though.

The mapping of status codes is probably not correct at this point in time.