• More cleaning up of commit stuff.
The major change is that we use the new Subversion::StatusListing class to inspect the state of what we want to copy. Subversion.status() now returns an object of this new class.

The old commit_status_helper is gone and replaced with commit_helper that uses Subversion::StatusListing as well. 

The options in commit.rb are no longer global. Since we refactored this to separate into commit_transaction it is no longer needed. 

The CommitWindow is now opened with TextMate::Process.run().