• Route all TM commands through ‘tmsvn_bootstrap.sh’ which asserts that we have svn and sets TM_SVN (if not set). It in turn calls ‘tmsvn.rb’ which launches the _actual_  ruby script to do the work.
So why all the indirection? 

‘tmsvn_bootstrap’ prevents us from having to put ‘require_cmd svn’  in each command in TM.

The actual scripts that do the work need to be launched in the background. Otherwise, the TM GUI locks up and we can't interact with it for the duration of the script. If we just launched the _actual_ script in background then any errors that occurred would not be represented which would be extremely confusing for users. So the bootstrap launches ‘tmsvn.rb’ in the background which throws alerts if the actual script that it launches runs into problems.