• Rework the revert command.
We actually lost a little bit of functionality here. The previous version only listed files that would actually be changed by the revert in the confirmation dialog, and it listed their pre-revert status.

I prefer the listing to show all files that will be passed to revert than trying to workout if they will be changed or not. Also, I prefer the cleaner dialog without the status codes, but I can add that back in if requested.