• Rework all of the diff commands.
diff.rb provides a wrapper over the DiffOperation class. It accepts a ‘revision’ argument where the '?' and ':' characters have special meaning. 

When the revision arg is '?', the revision chooser is called to select one revision.

When the revision arg is ':', the revision chooser is called to select two revisions (i.e. a range).

By default diff.rb sends the diff to STDOUT, but if the --send-to-mate switch is set the diff is written to a temp file and opened with mate.