• Use a REXML parse listener to create the status listing ruby object.
The REXML that ships with Ruby seems to be different on 10.4 and lacked some of the methods that this previously relied upon. Anyways, this is probably the better approach.

I made the decision to store the data as a network of hashes to make it easier to pass to tm_dialog (our ruby » plist converter can't convert arbitrary objects). It makes things a little looser, but is for the greater good.