Reworking of the update command…
This has been split out into 3 components: the actual execution of svn update, parsing of the output into a model, and rendering that result into html.

It has also been divorced from the `svn status` code. While there are similarities, trying to share the code made it pretty hard to follow.

I have also changed the strategy for styling the html. Each of the the style.css files now include common.css. This is opposed to including the css in view render.

I have also deleted the ‘Update Entire Project (Update Selected Files to Newest (HEAD))’ as it was excluded from the menu and duplicates the existing update command.