• Move command.
The user is asked to select a directory in the project as the destination for the move. If a single file is being moved (i.e. 0 or 1 files were selected in the project drawer when invoked) then the user is prompted to enter the name of the file at the new destination (the default value is the original name).