Add Subversion::esc helper function
All paths passed to the svn command line tool should go through this function. Presently the only thing it does is to append a ‘@’ to the filename if the filename already contains one. This is required because subversion cuts any suffix matching /@[^@]*$/ and treats that as shorthand for the revision argument (-r).

Ideally svn would have an option to disable this behavior.