Allow ‘@’ in filenames for Add, Commit, and Log
Doing a complete fix (for all commands) is very complicated as there is no ideal place to add this escaping.

Ideally we would escape at the level where we pass the paths to ‘svn’, but the current API simply provide an argument array that may contain things other than paths, and the commit window can also call ‘svn’ (e.g. for add and diff).

Given the decreased use of subversion and the rarity of ‘@’ in paths, I don’t think it’s worth investing the time to overhaul the bundle to better handle this issue.

Fixes #5.