With no selection we update entire project
Previously we would fallback to current document, when nothing was selected in the file browser.

This seemed consistent with the Add, Remove, and Revert commands, though for updating, we generally do want to batch update, and limiting a command to current file can easily be done using ⌃⌘R (select current document).